feat: update greeting feature and add a warning for node 22 bug
