+++ src/index.js
@@ -1,48 +1,13 @@
-import { DEFAULT_USERNAME_TEXT } from "./utils/constants.js";
-import { printMagentaText } from "./utils/get-color-coded-text.js";
+import { getUsername } from "./components/get-username.js";
+import { NODE_WARNING } from "./utils/constants.js";
+import { printBlueText, printMagentaText } from "./utils/get-color-coded-text.js";
 import { getGreetingText } from "./utils/get-text.js";
 
-export const APP_STATE = {
-    username: 'Username',
-} 
-
-function setUsername(){
-    const nameArgIndex = 2;
-    const nameArg = '--username';
-    const nameValueArgSeparator = '=';
-
-    console.log(process.argv[2], process.argv[3])
-
-    const argNameAndValueCurrent = process.argv[nameArgIndex];
-
-    if (!argNameAndValueCurrent) {
-        printMagentaText(DEFAULT_USERNAME_TEXT);
-
-        return;
-    }
-
-    const argNameCurrent = argNameAndValueCurrent.split(nameValueArgSeparator)[0];
-    const argValueCurrent = argNameAndValueCurrent.split(nameValueArgSeparator)[1];
-
-    if (
-        !argNameCurrent 
-        || !argValueCurrent 
-        ||argNameCurrent !== nameArg 
-    ) {
-        printMagentaText(DEFAULT_USERNAME_TEXT);
-    } else {
-        APP_STATE.username = argValueCurrent;
-    }
-
-    
-}
-
 function startApp(){
-    setUsername();
+    printBlueText(NODE_WARNING);
 
-    printMagentaText(getGreetingText(APP_STATE.username));
+    const userName = getUsername();
+    printMagentaText(getGreetingText(userName));
 } 
 
 startApp();
-
-console.log(APP_STATE.username, 'at the end');
