+++ src/utils/constants.js
@@ -4,4 +4,7 @@ export const INVALID_INPUT_TEXT = 'Invalid input.\n';
 
 export const PROMPT_TEXT = 'Please put in a command.\n';
 
-export const DEFAULT_USERNAME_TEXT = 'Ypu failed to provide your name in the correct format. Assigning a default username to you.\n'
+export const DEFAULT_USERNAME_TEXT = `You failed to provide your name in the correct format. Assigning a default username to you.\n`;
+
+export const NODE_WARNING = `If you're using Node 22 or higher, please start the app by 'npm run -- -- --username=your_name'
+due to the bug described on https://github.com/npm/cli/issues/7375\n\n`;
